Functions separated for stm32f1xx and stm32f4xx
